% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_city_canada.R
\name{every_city_plot}
\alias{every_city_plot}
\title{Generate every city canada card for city, returns path to image file}
\usage{
every_city_plot(city, file_path = NA)
}
\description{
Generate every city canada card for city, returns path to image file
}
