% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_city_canada.R
\name{not_tweeted_cities_list}
\alias{not_tweeted_cities_list}
\title{List of cities that have not yet been tweeted}
\usage{
not_tweeted_cities_list(tweeted_cities_path = file.path(here::here(),
  "data", "tweeted_cities"))
}
\description{
List of cities that have not yet been tweeted
}
