% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_city_canada.R
\name{send_every_city_tweet}
\alias{send_every_city_tweet}
\title{Tweet out every city canada card for random city}
\usage{
send_every_city_tweet(city = NA, media_feedback = TRUE,
  tweeted_cities_path = file.path(here::here(), "data",
  "tweeted_cities"))
}
\description{
Tweet out every city canada card for random city
}
