% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_city_canada.R
\name{tweeted_cities_list}
\alias{tweeted_cities_list}
\title{List of cities already tweeted}
\usage{
tweeted_cities_list(tweeted_cities_path = file.path(here::here(), "data",
  "tweeted_cities"))
}
\description{
List of cities already tweeted
}
